Fixes broken imports when AST is reused

+++ index.js
@@ -1,5 +1,7 @@
 const fs = require('fs');
 
+const isUrlRegExp = /(?:url|(?:-webkit-)?image-set)\(/i;
+
 const { NAMESPACE } = require('./config');
 
 const cache = new Map();
@@ -33,10 +35,7 @@ function pitch(remainingRequest, precedingRequest, data) {
     // if we previously detected that file didn't changed for this compilation
     // or if we already processed this file, just use cache
     } else if (cachedFile.useCache) {
-      this.loaders = this.loaders.filter(
-        (loader, index) => index <= this.loaderIndex
-      );
-
+      preventOngoingLoaders(this);
       return;
     }
   }
@@ -48,9 +47,7 @@ function pitch(remainingRequest, precedingRequest, data) {
   if (sameFileCompilationQueue.has(this.resourcePath)) {
     const queue = sameFileCompilationQueue.get(this.resourcePath);
 
-    this.loaders = this.loaders.filter(
-      (loader, index) => index <= this.loaderIndex
-    );
+    preventOngoingLoaders(this);
 
     queue.push(cb);
 
@@ -68,14 +65,12 @@ function pitch(remainingRequest, precedingRequest, data) {
         mtime,
         lastChecked: compilationStartTime
       });
-
-    // if new compilation but file didn't change, use old values
-    } else if (cachedFile.mtime === mtime) {
+      
+      // if new compilation but file didn't change, use old values
+    } else if (cachedFile.mtime.getTime() === mtime.getTime()) {
       cachedFile.useCache = true;
 
-      this.loaders = this.loaders.filter(
-        (loader, index) => index <= this.loaderIndex
-      );
+      preventOngoingLoaders(this);
     
     // if file changed since last compilation, force recompilation
     } else {
@@ -96,7 +91,14 @@ function loader(css, map, meta) {
   let result;
 
   if (cachedFile && cachedFile.useCache) {
-    result = [null, cachedFile.css, map, cachedFile.meta];
+    if (cachedFile.urls.size) {
+      restoreUrls(cachedFile.urls);
+    }
+    if (cachedFile.imports.length) {
+      restoreImports(cachedFile.meta.ast, cachedFile.imports);
+    }
+
+    result = [null, cachedFile.css, cachedFile.map, cachedFile.meta];
 
   // after file compiled, cache it, and mark that it's safe to use cached value
   } else {
@@ -104,20 +106,57 @@ function loader(css, map, meta) {
       useCache: true,
       css,
       map,
-      meta
+      meta,
+      imports: extractImports(meta.ast.root),
+      urls: extractUrls(meta.ast.root)
     });
 
     result = [null, css, map, meta];
   }
   
-  this.callback(...result);
-
   // if there are other calls to compilation of this file,
   // then we need to resolve them and clear queue
-  if (queue) {
+  if (queue && queue.length) {
     queue.forEach(cb => cb());
     sameFileCompilationQueue.delete(this.resourcePath);
   }
+  
+  this.callback(...result);
+}
+
+function restoreUrls(urls) {
+  urls.forEach((val, key, map) => {
+    key.value = val;
+  });
+}
+
+function extractUrls(ast) {
+  const urlsMap = new Map();
+  
+  ast.walkDecls(decl => {
+    if (!isUrlRegExp.test(decl.value)) {
+      return;
+    }
+
+    urlsMap.set(decl, decl.value);
+  });
+
+  return urlsMap;
+}
+
+function restoreImports(ast, imports) {
+  ast.root.nodes.unshift(...imports);
+  imports.forEach(item => item.parent = ast.root);
+}
+
+function extractImports(ast) {
+  const imports = [];
+
+  ast.walkAtRules('import', rule => {
+    imports.push(rule);
+  });
+
+  return imports;
 }
 
 /**
@@ -131,11 +170,21 @@ function findPlugin(loader) {
     : null;
   const plugin = parentCompiler
     ? parentCompiler.options.plugins.find(
-        p => p.NAMESPACE && p.NAMESPACE === NAMESPACE
-      )
+      p => p.NAMESPACE && p.NAMESPACE === NAMESPACE
+    )
     : this[NAMESPACE];
 
   return plugin;
+}
+
+function preventOngoingLoaders(currentLoader) {
+  debugger;
+  currentLoader.loaders.forEach((loader, index) => {
+    if (index <= currentLoader.loaderIndex) return;
+
+    loader.pitchExecuted = true;
+    loader.normalExecuted = true;
+  });
 }
 
 module.exports = loader;
